- add hilt to liquid module - set BaseBottomSheetDialog.kt with handling of autologout - ignore db schema files on git - move LockScreenViewModel.kt to ui package
