+++ .gitignore
@@ -30,3 +30,4 @@ fastlane/report.xml
 google-services.json
 output-metadata.json
 service.properties
+wallet/schemas/*/*.json
