+++ liquid-integration/src/main/java/org/dash/wallet/integration/liquid/dialog/SelectBuyDashDialog.kt
@@ -3,12 +3,14 @@ package org.dash.wallet.integration.liquid.dialog
 import android.content.Context
 import android.widget.ImageView
 import android.widget.LinearLayout
+import androidx.lifecycle.Lifecycle
 import org.dash.wallet.integration.liquid.listener.ValueSelectListener
 import com.google.android.material.bottomsheet.BottomSheetDialog
+import org.dash.wallet.common.ui.BaseBottomSheetDialog
 import org.dash.wallet.integration.liquid.R
 
 
-class SelectBuyDashDialog(val contexts: Context, private val listener: ValueSelectListener) : BottomSheetDialog(contexts) {
+class SelectBuyDashDialog(val contexts: Context, val lifecycle: Lifecycle, private val listener: ValueSelectListener) : BaseBottomSheetDialog(contexts, lifecycle) {
 
     init {
         setCancelable(true)
