+++ uphold-integration/src/main/java/org/dash/wallet/integration/uphold/ui/UpholdWithdrawalHelper.java
@@ -21,12 +21,9 @@ import android.app.ProgressDialog;
 import android.content.Intent;
 import android.net.Uri;
 
-import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.lifecycle.LifecycleOwner;
-import androidx.lifecycle.ViewModelStoreOwner;
 
-import org.dash.wallet.common.util.AlertDialogBuilder;
+import org.dash.wallet.common.ui.AlertDialogBuilder;
 import org.dash.wallet.integration.uphold.R;
 import org.dash.wallet.integration.uphold.data.UpholdApiException;
 import org.dash.wallet.integration.uphold.data.UpholdClient;
