+++ wallet/src/de/schildbach/wallet/ui/BaseBottomSheetDialogFragment.kt
@@ -27,7 +27,7 @@ import com.google.android.material.bottomsheet.BottomSheetDialog
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 import dagger.hilt.android.AndroidEntryPoint
 import de.schildbach.wallet_test.R
-import org.dash.wallet.common.LockScreenViewModel
+import org.dash.wallet.common.ui.LockScreenViewModel
 import org.dash.wallet.common.UserInteractionAwareCallback
 
 @AndroidEntryPoint
