+++ wallet/src/de/schildbach/wallet/ui/RestoreWalletFromSeedDialogFragment.java
@@ -38,7 +38,7 @@ import android.widget.TextView;
 
 import org.bitcoinj.crypto.MnemonicException;
 import org.bitcoinj.wallet.Wallet;
-import org.dash.wallet.common.util.AlertDialogBuilder;
+import org.dash.wallet.common.ui.AlertDialogBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
