+++ wallet/src/de/schildbach/wallet/ui/backup/RestoreFromFileActivity.kt
@@ -35,7 +35,7 @@ import de.schildbach.wallet.ui.widget.UpgradeWalletDisclaimerDialog
 import de.schildbach.wallet_test.R
 import org.bitcoinj.wallet.Wallet
 import org.dash.wallet.common.SecureActivity
-import org.dash.wallet.common.util.AlertDialogBuilder
+import org.dash.wallet.common.ui.AlertDialogBuilder
 
 
 @SuppressLint("Registered")
